+++ src/main/java/net/mcreator/bettertoolsandarmor/procedures/CrystalliteChestplateEmeraldProcedureProcedure.java
@@ -37,7 +37,7 @@ public class CrystalliteChestplateEmeraldProcedureProcedure {
 			if (entity.getData(BetterToolsModVariables.PLAYER_VARIABLES).crystallite_emerald_heal_timer == 0) {
 				if ((entity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) < (entity instanceof LivingEntity _livEnt ? _livEnt.getMaxHealth() : -1)) {
 					if (entity instanceof LivingEntity _entity)
-						_entity.setHealth((float) ((entity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) + 1));
+						_entity.setHealth((entity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) + 1);
 					if (world instanceof ServerLevel _level)
 						_level.sendParticles(ParticleTypes.HAPPY_VILLAGER, x, (y + 1), z, 5, 0.5, 0.75, 0.5, 0.0025);
 				}
