+++ src/main/java/net/mcreator/bettertoolsandarmor/procedures/CrystalliteHoeLapisProcedureProcedure.java
@@ -50,7 +50,7 @@ public class CrystalliteHoeLapisProcedureProcedure {
 				if (Math.random() < chance) {
 					if (world instanceof ServerLevel _level)
 						_level.addFreshEntity(new ExperienceOrb(_level, (x + Math.random()), (y + Math.random()), (z + Math.random()), Mth.nextInt(RandomSource.create(), 1,
-								(int) ((entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(Enchantments.FORTUNE)) + 1))));
+								(entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(Enchantments.FORTUNE)) + 1)));
 				}
 			}
 		}
