+++ src/main/java/net/mcreator/bettertoolsandarmor/procedures/ExperienceEnchProcedureProcedure.java
@@ -36,8 +36,8 @@ public class ExperienceEnchProcedureProcedure {
 		if ((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)
 				.getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.parse("better_tools:experience_ench")))) != 0) {
 			if (world instanceof ServerLevel _level)
-				_level.addFreshEntity(new ExperienceOrb(_level, x, y, z, (int) ((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)
-						.getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.parse("better_tools:experience_ench")))) * 5)));
+				_level.addFreshEntity(new ExperienceOrb(_level, x, y, z, (sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)
+						.getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.parse("better_tools:experience_ench")))) * 5));
 		}
 	}
 }
