+++ src/main/java/net/mcreator/bettertoolsandarmor/procedures/SugarClumpProcedureProcedure.java
@@ -14,8 +14,7 @@ public class SugarClumpProcedureProcedure {
 				_entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 100, 0));
 		} else if ((entity instanceof LivingEntity _livEnt && _livEnt.hasEffect(MobEffects.MOVEMENT_SPEED) ? _livEnt.getEffect(MobEffects.MOVEMENT_SPEED).getAmplifier() : 0) == 0) {
 			if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
-				_entity.addEffect(
-						new MobEffectInstance(MobEffects.MOVEMENT_SPEED, (int) ((entity instanceof LivingEntity _livEnt && _livEnt.hasEffect(MobEffects.MOVEMENT_SPEED) ? _livEnt.getEffect(MobEffects.MOVEMENT_SPEED).getDuration() : 0) + 100), 0));
+				_entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, (entity instanceof LivingEntity _livEnt && _livEnt.hasEffect(MobEffects.MOVEMENT_SPEED) ? _livEnt.getEffect(MobEffects.MOVEMENT_SPEED).getDuration() : 0) + 100, 0));
 		}
 	}
 }
